Iox #32 Mac OS fully supported

+++ iceoryx_utils/source/posix_wrapper/message_queue.cpp
@@ -165,7 +165,7 @@ cxx::expected<IpcChannelError> MessageQueue::destroy()
 cxx::expected<IpcChannelError> MessageQueue::send(const std::string& msg) const
 {
     const size_t messageSize = static_cast<size_t>(msg.size()) + NULL_TERMINATOR_SIZE;
-    if (messageSize > m_attributes.mq_msgsize)
+    if (messageSize > static_cast<size_t>(m_attributes.mq_msgsize))
     {
         return cxx::error<IpcChannelError>(IpcChannelError::MESSAGE_TOO_LONG);
     }
@@ -318,7 +318,7 @@ cxx::expected<std::string, IpcChannelError> MessageQueue::timedReceive(const uni
 cxx::expected<IpcChannelError> MessageQueue::timedSend(const std::string& msg, const units::Duration& timeout) const
 {
     const size_t messageSize = static_cast<size_t>(msg.size()) + NULL_TERMINATOR_SIZE;
-    if (messageSize > m_attributes.mq_msgsize)
+    if (messageSize > static_cast<size_t>(m_attributes.mq_msgsize))
     {
         std::cerr << "the message \"" << msg << "\" which should be sent to the message queue \"" << m_name
                   << "\" is too long" << std::endl;
